refactor: resolved one more TODO
