+++ TODO
@@ -4,7 +4,7 @@
 - (x) mention in readme/package.json
 - (x) design a multifunctional API
 - (x) draw lottie animation (re-use existing one? for KAV?)
-- test StickyFooter on chat examples
-- test how keyboard changes size
-- use common TextInput component - same as usual KeyboardAwareScrollView example
+- (x) test StickyFooter on chat examples
+- (x) test how keyboard changes size
+- (x) use common TextInput component - same as usual KeyboardAwareScrollView example
 - port example to Fabric
